+++ internal/app/singularity/oci_pause_linux.go
@@ -6,68 +6,39 @@
 package singularity
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
+	"syscall"
 
-	"github.com/sylabs/singularity/pkg/ociruntime"
-	"github.com/sylabs/singularity/pkg/util/unix"
+	"github.com/sylabs/singularity/pkg/sylog"
 )
 
-// OciPauseResume pauses/resumes processes in a container
-func OciPauseResume(containerID string, pause bool) error {
-	state, err := getState(containerID)
-	if err != nil {
-		return err
+// OciPause pauses processes in a container
+func OciPause(containerID string) error {
+	runcArgs := []string{
+		"--root=" + OciStateDir,
+		"pause",
+		containerID,
 	}
 
-	if state.ControlSocket == "" {
-		return fmt.Errorf("can't find control socket")
+	sylog.Debugf("Calling runc with args %v", runcArgs)
+	if err := syscall.Exec(runc, runcArgs, []string{}); err != nil {
+		return fmt.Errorf("while calling runc: %w", err)
 	}
 
-	if pause && state.Status != ociruntime.Running {
-		return fmt.Errorf("container %s is not running", containerID)
-	} else if !pause && state.Status != ociruntime.Paused {
-		return fmt.Errorf("container %s is not paused", containerID)
+	return nil
+}
+
+// OciResume pauses processes in a container
+func OciResume(containerID string) error {
+	runcArgs := []string{
+		"--root=" + OciStateDir,
+		"resume",
+		containerID,
 	}
 
-	ctrl := &ociruntime.Control{}
-	if pause {
-		ctrl.Pause = true
-	} else {
-		ctrl.Resume = true
-	}
-
-	c, err := unix.Dial(state.ControlSocket)
-	if err != nil {
-		return fmt.Errorf("failed to connect to control socket")
-	}
-	defer c.Close()
-
-	enc := json.NewEncoder(c)
-	if enc == nil {
-		return fmt.Errorf("cannot instantiate new JSON encoder")
-	}
-
-	if err := enc.Encode(ctrl); err != nil {
-		return err
-	}
-
-	// wait runtime close socket connection for ACK
-	d := make([]byte, 1)
-	if _, err := c.Read(d); err != io.EOF {
-		return err
-	}
-
-	// check status
-	state, err = getState(containerID)
-	if err != nil {
-		return err
-	}
-	if pause && state.Status != ociruntime.Paused {
-		return fmt.Errorf("bad status %s returned instead of paused", state.Status)
-	} else if !pause && state.Status != ociruntime.Running {
-		return fmt.Errorf("bad status %s returned instead of running", state.Status)
+	sylog.Debugf("Calling runc with args %v", runcArgs)
+	if err := syscall.Exec(runc, runcArgs, []string{}); err != nil {
+		return fmt.Errorf("while calling runc: %w", err)
 	}
 
 	return nil