+++ internal/app/singularity/oci_start_linux.go
@@ -6,51 +6,23 @@
 package singularity
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
+	"syscall"
 
-	"github.com/sylabs/singularity/pkg/ociruntime"
-	"github.com/sylabs/singularity/pkg/util/unix"
+	"github.com/sylabs/singularity/pkg/sylog"
 )
 
 // OciStart starts a previously create container
 func OciStart(containerID string) error {
-	state, err := getState(containerID)
-	if err != nil {
-		return err
+	runcArgs := []string{
+		"--root=" + OciStateDir,
+		"start",
+		containerID,
 	}
 
-	if state.Status != ociruntime.Created {
-		return fmt.Errorf("cannot start '%s', the state of the container must be %s", containerID, ociruntime.Created)
-	}
-
-	if state.ControlSocket == "" {
-		return fmt.Errorf("can't find control socket")
-	}
-
-	ctrl := &ociruntime.Control{}
-	ctrl.StartContainer = true
-
-	c, err := unix.Dial(state.ControlSocket)
-	if err != nil {
-		return fmt.Errorf("failed to connect to control socket")
-	}
-	defer c.Close()
-
-	enc := json.NewEncoder(c)
-	if enc == nil {
-		return fmt.Errorf("cannot instantiate new JSON encoder")
-	}
-
-	if err := enc.Encode(ctrl); err != nil {
-		return err
-	}
-
-	// wait runtime close socket connection for ACK
-	d := make([]byte, 1)
-	if _, err := c.Read(d); err != io.EOF {
-		return err
+	sylog.Debugf("Calling runc with args %v", runcArgs)
+	if err := syscall.Exec(runc, runcArgs, []string{}); err != nil {
+		return fmt.Errorf("while calling runc: %w", err)
 	}
 
 	return nil