+++ internal/app/singularity/oci_state_linux.go
@@ -6,32 +6,24 @@
 package singularity
 
 import (
-	"encoding/json"
 	"fmt"
+	"syscall"
 
-	"github.com/sylabs/singularity/pkg/util/unix"
+	"github.com/sylabs/singularity/pkg/sylog"
 )
 
 // OciState query container state
 func OciState(containerID string, args *OciArgs) error {
-	// query instance files and returns state
-	state, err := getState(containerID)
-	if err != nil {
-		return err
+	runcArgs := []string{
+		"--root=" + OciStateDir,
+		"state",
+		containerID,
 	}
-	if args.SyncSocketPath != "" {
-		data, err := json.Marshal(state)
-		if err != nil {
-			return fmt.Errorf("failed to marshal state data: %s", err)
-		} else if err := unix.WriteSocket(args.SyncSocketPath, data); err != nil {
-			return err
-		}
-	} else {
-		c, err := json.MarshalIndent(state, "", "\t")
-		if err != nil {
-			return err
-		}
-		fmt.Println(string(c))
+
+	sylog.Debugf("Calling runc with args %v", runcArgs)
+	if err := syscall.Exec(runc, runcArgs, []string{}); err != nil {
+		return fmt.Errorf("while calling runc: %w", err)
 	}
+
 	return nil
 }