+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - systemd is now supported and used as the default cgroups manager. Set
   `systemd cgroups = no` in `singularity.conf` to manage cgroups directly via
   the cgroupfs.
+- The `singularity oci` command group now uses `runc` to manage containers.
+- The `singularity oci` flags `--sync-socket`, `--empty-process`, and
+  `--timeout` have been removed.
 
 ### New features / functionalities